+++ src/main/java/net/pms/network/mediaserver/jupnp/UmsServerUpnpServiceConfiguration.java
@@ -17,12 +17,12 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package net.pms.network.mediaserver.cling;
+package net.pms.network.mediaserver.jupnp;
 
 import net.pms.network.mediaserver.MediaServer;
-import net.pms.network.mediaserver.cling.transport.impl.JdkHttpServerStreamServer;
-import net.pms.network.mediaserver.cling.transport.impl.NettyStreamServer;
-import net.pms.network.mediaserver.cling.transport.impl.UmsStreamServerConfiguration;
+import net.pms.network.mediaserver.jupnp.transport.impl.JdkHttpServerStreamServer;
+import net.pms.network.mediaserver.jupnp.transport.impl.NettyStreamServer;
+import net.pms.network.mediaserver.jupnp.transport.impl.UmsStreamServerConfiguration;
 import org.jupnp.transport.spi.NetworkAddressFactory;
 import org.jupnp.transport.spi.StreamServer;
