+++ src/main/java/net/pms/network/mediaserver/jupnp/support/contentdirectory/UMSContentDirectoryService.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package net.pms.network.mediaserver.cling.support.contentdirectory;
+package net.pms.network.mediaserver.jupnp.support.contentdirectory;
 
 import org.jupnp.support.contentdirectory.AbstractContentDirectoryService;
 import org.jupnp.support.contentdirectory.ContentDirectoryErrorCode;
