+++ src/main/java/net/pms/network/mediaserver/jupnp/transport/impl/UmsStreamServerConfiguration.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package net.pms.network.mediaserver.cling.transport.impl;
+package net.pms.network.mediaserver.jupnp.transport.impl;
 
 import org.jupnp.transport.spi.StreamServerConfiguration;
