+++ src/main/java/net/pms/network/mediaserver/socketchannelserver/RequestHandler.java
@@ -133,7 +133,7 @@ public class RequestHandler implements Runnable {
 					request.setMediaRenderer(renderer);
 				}
 				if (headerLine.toUpperCase().startsWith("USER-AGENT")) {
-					// Is the request from our own Cling service, i.e. self-originating?
+					// Is the request from our own JUPnP service, i.e. self-originating?
 					if (isSelf && headerLine.contains("UMS/")) {
 						//LOGGER.trace("Ignoring self-originating request from {}:{}", ia, remoteAddress.getPort());
 						return;
